+++ src/entity/builtin_entity.rs
@@ -57,7 +57,10 @@ enum_kind!(
         Percentage,
         MusicAlbum,
         MusicArtist,
-        MusicTrack
+        MusicTrack,
+        City,
+        Country,
+        Region
     ]
 );
 
@@ -106,6 +109,9 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => "snips/musicAlbum",
             BuiltinEntityKind::MusicArtist => "snips/musicArtist",
             BuiltinEntityKind::MusicTrack => "snips/musicTrack",
+            BuiltinEntityKind::City => "snips/city",
+            BuiltinEntityKind::Country => "snips/country",
+            BuiltinEntityKind::Region => "snips/region",
         }
     }
 
@@ -135,6 +141,9 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => "Matches a music album",
             BuiltinEntityKind::MusicArtist => "Matches a music artist",
             BuiltinEntityKind::MusicTrack => "Matches a music track",
+            BuiltinEntityKind::City => "Matches main local and world cities",
+            BuiltinEntityKind::Country => "Matches countries around the world",
+            BuiltinEntityKind::Region => "Matches local administrative regions",
         }
     }
 }
@@ -148,8 +157,8 @@ impl BuiltinEntityKind {
             Language::FR => self.fr_examples(),
             Language::JA => self.ja_examples(),
             Language::IT => self.it_examples(),
-            Language::PT_PT => self.pt_examples(),
-            Language::PT_BR => self.pt_examples(),
+            Language::PT_PT => self.pt_pt_examples(),
+            Language::PT_BR => self.pt_br_examples(),
             Language::KO => self.ko_examples(),
         }
     }
@@ -197,6 +206,19 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => &["Discovery"],
             BuiltinEntityKind::MusicArtist => &["Daft Punk"],
             BuiltinEntityKind::MusicTrack => &["Harder Better Faster Stronger"],
+            BuiltinEntityKind::City => &[
+                "Berlin",
+                "Essen",
+                "Zürich",
+                "Paris",
+            ],
+            BuiltinEntityKind::Country => &[
+                "Frankreich",
+            ],
+            BuiltinEntityKind::Region => &[
+                "Bayern",
+                "Tirol",
+            ],
         }
     }
 
@@ -274,6 +296,19 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => &["Discovery"],
             BuiltinEntityKind::MusicArtist => &["Daft Punk"],
             BuiltinEntityKind::MusicTrack => &["Harder Better Faster Stronger"],
+            BuiltinEntityKind::City => &[
+                "San Francisco",
+                "Los Angeles",
+                "Beijing",
+                "Paris",
+            ],
+            BuiltinEntityKind::Country => &[
+                "France",
+            ],
+            BuiltinEntityKind::Region => &[
+                "California",
+                "Washington",
+            ],
         }
     }
 
@@ -360,6 +395,19 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => &["Discovery"],
             BuiltinEntityKind::MusicArtist => &["Daft Punk"],
             BuiltinEntityKind::MusicTrack => &["Harder Better Faster Stronger"],
+            BuiltinEntityKind::City => &[
+                "Madrid",
+                "Barcelona",
+                "Bilbao",
+                "Paris",
+            ],
+            BuiltinEntityKind::Country => &[
+                "Francia",
+            ],
+            BuiltinEntityKind::Region => &[
+                "Andalusia",
+                "Catalonia",
+            ],
         }
     }
 
@@ -418,6 +466,21 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => &["Discovery"],
             BuiltinEntityKind::MusicArtist => &["Daft Punk"],
             BuiltinEntityKind::MusicTrack => &["Harder Better Faster Stronger"],
+            BuiltinEntityKind::City => &[
+                "Paris",
+                "Brest",
+                "Bruxelles",
+                "Pékin",
+                "Londres",
+            ],
+            BuiltinEntityKind::Country => &[
+                "France",
+            ],
+            BuiltinEntityKind::Region => &[
+                "Bretagne",
+                "Corse",
+                "Province de Liège",
+            ],
         }
     }
 
@@ -506,10 +569,22 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => &["Discovery"],
             BuiltinEntityKind::MusicArtist => &["Daft Punk"],
             BuiltinEntityKind::MusicTrack => &["Harder Better Faster Stronger"],
+            BuiltinEntityKind::City => &[
+                "San Francisco",
+                "Roma",
+                "Agrigente",
+            ],
+            BuiltinEntityKind::Country => &[
+                "Italia",
+            ],
+            BuiltinEntityKind::Region => &[
+                "Sardinia",
+                "Sicilia",
+            ],
         }
     }
 
-    fn pt_examples(&self) -> &'static [&'static str] {
+    fn pt_br_examples(&self) -> &'static [&'static str] {
         match *self {
             BuiltinEntityKind::AmountOfMoney => &[
                 "10$",
@@ -520,6 +595,16 @@ impl BuiltinEntityKind {
                 "dez libras esterlinas",
                 "845584 francos suíços",
             ],
+            // Datetime subtypes not supported for this language,
+            // cf. BuiltinEntityKind.supported_languages(),
+            // but how to make this function aware of that?
+            BuiltinEntityKind::Datetime => &[
+                "hoje",
+            ],
+            BuiltinEntityKind::Date => &[],
+            BuiltinEntityKind::Time => &[],
+            BuiltinEntityKind::DatePeriod => &[],
+            BuiltinEntityKind::TimePeriod => &[],
             BuiltinEntityKind::Duration => &[
                 "1 hora",
                 "3 meses",
@@ -541,22 +626,89 @@ impl BuiltinEntityKind {
                 "quatro graus centígrados",
                 "-459,67 °F",
             ],
-            BuiltinEntityKind::Datetime => &[
-                "hoje",
-            ],
-            // Datetime subtypes not supported for this language,
-            // cf. BuiltinEntityKind.supported_languages(),
-            // but how to make this function aware of that?
-            BuiltinEntityKind::Date => &[],
-            BuiltinEntityKind::Time => &[],
-            BuiltinEntityKind::DatePeriod => &[],
-            BuiltinEntityKind::TimePeriod => &[],
             BuiltinEntityKind::Percentage => &[
                 "25%",
             ],
             BuiltinEntityKind::MusicAlbum => &["Discovery"],
             BuiltinEntityKind::MusicArtist => &["Daft Punk"],
             BuiltinEntityKind::MusicTrack => &["Harder Better Faster Stronger"],
+            BuiltinEntityKind::City => &[
+                "São Paulo",
+                "Rio",
+                "Los Angeles",
+                "Paris",
+            ],
+            BuiltinEntityKind::Country => &[
+                "Brasil",
+            ],
+            BuiltinEntityKind::Region => &[
+                "Bahia",
+                "Amazonas",
+            ],
+        }
+    }
+
+    fn pt_pt_examples(&self) -> &'static [&'static str] {
+        match *self {
+            BuiltinEntityKind::AmountOfMoney => &[
+                "10$",
+                "15€",
+                "cinco euros",
+                "16,65 €",
+                "dois euros e cinco centavos",
+                "dez libras esterlinas",
+                "845584 francos suíços",
+            ],
+            // Datetime subtypes not supported for this language,
+            // cf. BuiltinEntityKind.supported_languages(),
+            // but how to make this function aware of that?
+            BuiltinEntityKind::Datetime => &[
+                "hoje",
+            ],
+            BuiltinEntityKind::Date => &[],
+            BuiltinEntityKind::Time => &[],
+            BuiltinEntityKind::DatePeriod => &[],
+            BuiltinEntityKind::TimePeriod => &[],
+            BuiltinEntityKind::Duration => &[
+                "1 hora",
+                "3 meses",
+                "dez minutos",
+                "meia hora",
+                "oito anos e dois semanas",
+                "um ano quatro semanas e tres horas",
+            ],
+            BuiltinEntityKind::Number => &[
+                "2001",
+            ],
+            BuiltinEntityKind::Ordinal => &[
+                "primeira",
+            ],
+            BuiltinEntityKind::Temperature => &[
+                "70 graus kelvin",
+                "3°C",
+                "dez graus",
+                "quatro graus centígrados",
+                "-459,67 °F",
+            ],
+            BuiltinEntityKind::Percentage => &[
+                "25%",
+            ],
+            BuiltinEntityKind::MusicAlbum => &["Discovery"],
+            BuiltinEntityKind::MusicArtist => &["Daft Punk"],
+            BuiltinEntityKind::MusicTrack => &["Harder Better Faster Stronger"],
+            BuiltinEntityKind::City => &[
+                "Liboa",
+                "Porto",
+                "Amadora",
+            ],
+            BuiltinEntityKind::Country => &[
+                "Portugal",
+                "Espanha",
+            ],
+            BuiltinEntityKind::Region => &[
+                "Norte",
+                "Alentejo",
+            ],
         }
     }
 
@@ -584,6 +736,18 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => &["Discovery"],
             BuiltinEntityKind::MusicArtist => &["Daft Punk"],
             BuiltinEntityKind::MusicTrack => &["Harder Better Faster Stronger"],
+            BuiltinEntityKind::City => &[
+                "パリ",
+                "東京",
+                "京都",
+            ],
+            BuiltinEntityKind::Country => &[
+                "日本",
+            ],
+            BuiltinEntityKind::Region => &[
+                "北海道",
+                "関東地方",
+            ],
         }
     }
 
@@ -606,6 +770,18 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => &["Discovery"],
             BuiltinEntityKind::MusicArtist => &["Daft Punk"],
             BuiltinEntityKind::MusicTrack => &["Harder Better Faster Stronger"],
+            BuiltinEntityKind::City => &[
+                "서울",
+                "파리",
+                "부산광역시",
+            ],
+            BuiltinEntityKind::Country => &[
+                "프랑스",
+            ],
+            BuiltinEntityKind::Region => &[
+                "충청도",
+                "강원도",
+            ],
         }
     }
 }
@@ -706,6 +882,21 @@ impl BuiltinEntityKind {
                     value: "Harder Better Faster Stronger".to_string(),
                 })])
             }
+            BuiltinEntityKind::City => {
+                serde_json::to_string_pretty(&vec![SlotValue::City(StringValue {
+                    value: "Paris".to_string(),
+                })])
+            }
+            BuiltinEntityKind::Country => {
+                serde_json::to_string_pretty(&vec![SlotValue::Country(StringValue {
+                    value: "France".to_string(),
+                })])
+            }
+            BuiltinEntityKind::Region => {
+                serde_json::to_string_pretty(&vec![SlotValue::Region(StringValue {
+                    value: "California".to_string(),
+                })])
+            }
         }
         .unwrap()
     }
@@ -797,6 +988,36 @@ impl BuiltinEntityKind {
             BuiltinEntityKind::MusicAlbum => &[Language::EN, Language::FR],
             BuiltinEntityKind::MusicArtist => &[Language::EN, Language::FR],
             BuiltinEntityKind::MusicTrack => &[Language::EN, Language::FR],
+            BuiltinEntityKind::City => &[
+                Language::DE,
+                Language::EN,
+                Language::ES,
+                Language::FR,
+                Language::IT,
+                Language::JA,
+                Language::PT_BR,
+                Language::PT_PT,
+            ],
+            BuiltinEntityKind::Country => &[
+                Language::DE,
+                Language::EN,
+                Language::ES,
+                Language::FR,
+                Language::IT,
+                Language::JA,
+                Language::PT_BR,
+                Language::PT_PT,
+            ],
+            BuiltinEntityKind::Region => &[
+                Language::DE,
+                Language::EN,
+                Language::ES,
+                Language::FR,
+                Language::IT,
+                Language::JA,
+                Language::PT_BR,
+                Language::PT_PT,
+            ],
         }
     }